Accessories done

+++ app/src/main/java/com/example/nikhil/product/Properties.java
@@ -9,11 +9,11 @@ public class Properties {
     public String code;
     public String description;
     public int pkg;
-    public int mrp;
+    public float mrp;
     public int image;
 
 
-    public Properties(String code_r, String description_r, int pkg_r, int mrp_r, int image_r){
+    public Properties(String code_r, String description_r, int pkg_r, float mrp_r, int image_r){
 
         code = code_r;
         description = description_r;
